pyproject.toml
[tool.poetry]
name = "botus_receptus"
version = "0.0.1a"
description = ""
authors = ["Bryan Forbes <bryan@reigndropsfall.net>"]
license = "BSD-3-Clause"
packages = [
	{ include = "botus_receptus", from = "src" },
	{ include = "asyncpg-stubs", from = "src" },
	{ include = "uvloop-stubs", from = "src" },
]

include = [
	"botus_receptus/py.typed",
	"asyncpg-stubs/**/*.pyi",
	"uvloop-stubs/**/*.pyi"
]

[tool.poetry.dependencies]
python = "^3.7"
aioitertools = "^0.5.0"
"async-timeout" = ">=3.0.0"
attrs = ">=18.2.0"
click = ">=6.7"
"discord.py" = {rev = "07db03a9ab22dc4fee959667900190ce37a552b1", git = "https://github.com/bryanforbes/discord.py.git", extras = ["typings"]}
"typing-extensions" = ">=3.7.4"
toml = "^0.10.0"
pendulum = {git = "https://github.com/bryanforbes/pendulum.git", branch = "update-for-pep-561"}
asyncpg = {version = ">=0.18.2", optional = true}
gino = {version = "~0.8.1", optional = true}
sqlalchemy = {version = "~1.2.18", optional = true}
gino-stubs = {rev = "3f1a3f26aa5a56d9fc58859d7ff140ca6cefe2fb", git = "https://github.com/bryanforbes/gino-stubs.git", optional = true}
uvloop = {version = "~0.14.0", optional = true}
more-itertools = {version = "~8.0.0", optional = true}

[tool.poetry.extras]
db = ["asyncpg"]
gino = ["gino", "sqlalchemy"]
gino_typed = ["gino", "sqlalchemy", "gino-stubs"]
more_itertools = ["more-itertools"]
uvloop = ["uvloop"]

[tool.poetry.dev-dependencies]
asynctest = "~0.13.0"
black = "^19.3b0"
flake8 = "^3.6"
flake8-bugbear = "^19.3"
flake8-pyi = "^19.2"
mypy = "^0.750"
pre-commit = "^1.12"
pytest = "^5.0.0"
pytest-asyncio = "^0.10.0"
pytest-cov = "^2.6"
pytest-mock = "^1.10"
pytest-sugar = "^0.9.1"
pytest-xdist = "^1.24"
isort = "^4.3.21"

[tool.black]
line-length = 88
target-version = ["py37"]
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | alembic
  | htmlcov
)/
'''

[tool.isort]
known_standard_library = ["typing_extensions"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=1.0.0b6"]
build-backend = "poetry.masonry.api"
